requirements.txt
asyncio
py-cord
aiosqlite
python-dotenv
random
datetime
io